[5.1] DarkMode selector (#43310)
On updated versions of Joomla there is no value set for the colorscheme and the only way to enable the colorscheme switch is to open the atum template style and then save it.

This PR is (hopefully) sql that will run on the next patch release that will add a default value for the colorScheme so that the switch is always available. Note that if a user has already set a value then nothing is changed.

I don't have a current postgres server so am relying on others to test that carefully.