Avoid setting an explicit session ID via GET args.
This is considered a failing metric in automated PCI scans under the
"session hijacking" category and thus should be avoided.

PHP 4.3 introduced the "session.use_only_cookies" PHP configuration
option which meant that passing in a session ID via GET/POST variables
can be disabled. The code in Joomla should at very least honour this
setting.

Alternatively, if no good reason for this code exists, it should be
removed entirely.