Fail in case of SMTP problems with a clear message
If the array `mail[]` is empty width length 0, the line
	`cy.wrap(mails).should('have.lengthOf', 1);`
does not cause the test to stop in this case as the Cypress
assertion inside the `.then()` has an asynchronous nature.
The effect is that the test run continues and the following access
to the mail fails with with a somehow incomprehensible message e.g.
	`TypeError: Cannot read properties of undefined (reading 'body')`

Replaced it with an `expect` statement, which immediately triggers an error,
and added an individual message that the SNMP configuration may need to be checked.

Added two more places.