[5.2] TinyMCE Accordion
Since Joomla 5.0 we have shipped a version of tinyMCE that includes a plugin called accordion. However we never exposed it to our plugin so it was not available to use.

This PR remedies that.
New installs will see the accordion icon on the toolbar with the list icons
For existing installs you will need to EITHER drag the icon into the toolbar in the plugin configuration OR if you havent changed the toolbar you can use the clear and the the use advanced preset buttons.

Technically tinymce have misnamed the the plugin as it should not be confused with other accordion functionality such as the js based solution that can be enabled with bootstrap.

This plugin implements the native html details/summary disclosure element which requires no javascript and is accessible out of the box.

To test
As this requires a js change either apply the pr and run `nmp ci` or use one of the prebuilt packages

If it is a new install then the icon will now be able in the toolbar. If it is an upgrade then you will need to use the tinymce plugin editor to add it (as described above)