[cypress] Updated to joomla-cypress 1.1.0
The joomla-cypress NPM package used has been updated to the latest version in order to be able to use one version in all active joomla-cms branches.
Other packages were also updated to the latest versions with the `npm update`.
The overwritten Cypress commands for faster login with session have been deleted, as they are already included in the new joomla-cypress version.