2nd ESM Update tests/System/README.md
With the https://github.com/joomla/joomla-cms/pull/43702 ESM documentation update, the small problem has arisen that `npx` uses `cypress.config.js` in error messages, although `cypress.config.mjs` is used. This was 'fixed' in using `npm`.

Unfortunately, it turned out that:
* npx cypress run --spec
is not replaceable by:
* npm run cypress:run --spec
therefore changed back.

A short note has been added to the Troubleshooting section.

It was also discovered that the environment variable in the `sudo` call was set in the wrong order, which has been corrected.