[5.2] Content Category Save to Menu
This PR adds two new menu items to the save dropdown menu on the toollbar allowing you to save a categorey directly to a menu item - either as a blog or a list just as you can directly save an article to a menu item.

This is currently restricted to categories in the com_content component. It could be expanded to other components but I didn't see the use case

Signed-off-by: BrianTeeman <brian@teeman.net>