[5.2] Custom Reply description
_new language string so has to be in 5.2_

There is an option for com_contact called Custom Reply. All it does is disable the sending of the contact from input but it could be used to allow a plugin to integrate with other systems and send the message that way.

I doubt it is used at all and its really a legacy from 1.6 when you could customise the reply directly but that functionality was removed.

This PR just adds a description to the field so if anyone was to look at it they wont be completely confused about tis purpose.

Signed-off-by: BrianTeeman <brian@teeman.net>