Adopt System Tests for new mod_community_info
The API test 'can deliver a list of administrator modules' originally
checked that the first entry in the list was `mod_sampledata`.
With the introduction of the new `mod_community_info` module,
this module now appears as the first entry.

Therefore, the test has been modified to check that `mod_sampledata`
exists somewhere in the list, rather than being the first entry.

All tests passed in the System Tests for 5.2dev after this commit,
both with and without PR 43753.