[cypress] Enable database usage with Unix Sockets
Check if db_host is a Unix socket by verifying the "unix:/" prefix.
JavaScript drivers does not handle this prefix, it needs to be stripped.
This approach maintains compatibility with the PHP drivers and allows
to have Cypress.config.mjs working for installation step (PHP driver)
and custom database commands (JavaScript driver).