[5.2] Debug plugin Remove unused code
This simple PR removes commented code which was intended to be an info link in the debug plugin. From the code we can see the intention was for there to be an info page about the debug plugin. As we don't have any compulsory content this can never work so might as well remove the code as it cant be implemented

code review as the code has never been used and is commented out.