[5.4] Media Delete Confirmation message
This PR introduces a dedicated delete confirmation message for com_media instead of the global delete message.

The reason is to further clarify/warn that deleting a folder will delete all the contents of that folder. Unline the generic message which just talks about deleting the selected item

To test you will need to either use a prebuilt package from this PR or apply the PR and then run npm i

Ideally I wanted to have a separate confirmation message for items and folders but this is not possible without significant rewrite and as no one has shown interest in that since the issue was first raised #29564 this PR is a small improvement.

Signed-off-by: BrianTeeman <brian@teeman.net>