[5.4] Blog Sample Data Module
When you install the blog sample data an additional module is added to the dashboard "Site Information" - you might need to refresh the admin dashboard page after installing the sample data to see it.

This module should be configured to be displayed the same as the other dashboard modules. This PR fixes a semantic error where the wrong heading level was set. IT should be h2 not h3. Bizaarely h3 is visually bigger than h2

For consistency with the other admin dashboard nmodules it also sets the bootstrap size to 12 from 0 but I dont think this has any visible impact.

Signed-off-by: BrianTeeman <brian@teeman.net>