[5.4] Remove duplicate string
This PR removes a duplicate loading of the language string JFIELD_MEDIA_LAZY_LABEL

Additionally I have alpha sorted the strings. Should make it easier in the future if a string is added to prevent it being added twice.

Sorry I have no idea where this is used so testing by code review or let me know where it is used and I can update the test instructions

Signed-off-by: BrianTeeman <brian@teeman.net>