[5.3] Fix fields menu items in preset
Missing view in the list fields menu item

Create a custom admin menu and import the alternate preset
In the menu manager check the User Fields and the Article Fields menu items. The Menu Item Type is empty
In the menu manager check the User Field Groups and the Article Field groups menu items.
The Menu Item Type is correct

Delete the custom admin menu And repeat the steps above
The Menu Item Type is now correct

[note the context is empty - this is fixed by 46052]

Signed-off-by: BrianTeeman <brian@teeman.net>